Avoid rewrite error when an update of a pointer with range bounds is in a macro
