+++ clang/include/clang/3C/AVarBoundsInfo.h
@@ -259,6 +259,8 @@ public:
   // to the bounds key.
   bool canInferRangeBounds(BoundsKey BK);
 
+  void markIneligibleForRangeBounds(BoundsKey BK);
+
   // Check if the bounds keys will need to be rewritten with range bounds. This
   // is true for bounds keys that are subject to pointer arithmetic but
   // otherwise have inferred bounds.
