+++ clang/lib/3C/ConstraintBuilder.cpp
@@ -10,6 +10,7 @@
 //===----------------------------------------------------------------------===//
 
 #include "clang/3C/ConstraintBuilder.h"
+#include "clang/3C/BasePointerAssignment.h"
 #include "clang/3C/3CGlobalOptions.h"
 #include "clang/3C/3CStats.h"
 #include "clang/3C/ArrayBoundsInferenceConsumer.h"
@@ -639,10 +640,12 @@ void VariableAdderConsumer::HandleTranslationUnit(ASTContext &C) {
   }
 
   VariableAdderVisitor VAV = VariableAdderVisitor(&C, Info);
+  BasePointerAssignmentFinder BPF = BasePointerAssignmentFinder(&C, Info);
   TranslationUnitDecl *TUD = C.getTranslationUnitDecl();
   // Collect Variables.
   for (const auto &D : TUD->decls()) {
     VAV.TraverseDecl(D);
+    BPF.TraverseDecl(D);
   }
 
   if (_3COpts.Verbose)
